requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.1
deepface==0.0.90
distlib==0.3.7
filelock==3.12.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
gast==0.5.4
gdown==5.1.0
google-pasta==0.2.0
grpcio==1.62.2
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mtcnn==0.1.1
namex==0.0.8
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==3.10.0
protobuf==4.25.3
PyAudio==0.2.14
Pygments==2.17.2
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytz==2024.1
pywin32==306
requests==2.31.0
retina-face==0.0.17
rich==13.7.1
six==1.16.0
soupsieve==2.5
SpeechRecognition==3.10.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.16.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
